pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "crescendo-interp"
version = "0.1.0"
description = "Mechanistic interpretability of multi-turn (Crescendo) jailbreak attacks"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "requests",
    "python-dotenv",
    "tqdm",
    "numpy",
]

[project.optional-dependencies]
pipelines = [
    "jaxtyping",
    "litellm",
    "datasets",
    "jsonlines",
    "matplotlib",
]
notebooks = [
    "jupyter",
    "plotly",
]

[tool.setuptools.packages.find]
where = ["src"]
